+++ src/main/java/com/norab/role/PlaysId.java
@@ -1,6 +1,4 @@
 package com.norab.role;
 
-public record PlaysId(Long id) {
+public record PlaysId(Integer id) {
 }
-
-
