+++ src/main/resources/db/migration/V4__PhotosTable.sql
@@ -1,19 +1,19 @@
 CREATE TABLE photos (
-    photo_id BIGSERIAL PRIMARY KEY,
+    photo_id SERIAL PRIMARY KEY,
     url TEXT,
-    movie_id BIGINT NULL,
-    actor_id BIGINT NULL,
-    role_id BIGINT NULL,
+    movie_id INTEGER NULL,
+    actor_id INTEGER NULL,
+    role_id  INTEGER NULL,
 		CONSTRAINT fk_photos_movie_id
-		FOREIGN KEY (movie_id)
-		REFERENCES movies (movie_id)
-		ON DELETE SET NULL,
+			FOREIGN KEY (movie_id)
+			REFERENCES movies (movie_id)
+			ON DELETE SET NULL,
 		CONSTRAINT fk_photos_actor_id
-		FOREIGN KEY (actor_id)
-		REFERENCES actors (actor_id)
-		ON DELETE SET NULL,
+			FOREIGN KEY (actor_id)
+			REFERENCES actors (actor_id)
+			ON DELETE SET NULL,
 		CONSTRAINT fk_photos_role_id
-		FOREIGN KEY (role_id)
-		REFERENCES plays (role_id)
-		ON DELETE SET NULL
+			FOREIGN KEY (role_id)
+			REFERENCES plays (role_id)
+			ON DELETE SET NULL
 );
